% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_alloc_NP.R
\name{fig_allocNP}
\alias{fig_allocNP}
\title{GRAS method}
\usage{
fig_allocNP(vec, targetValue)
}
\arguments{
\item{vec}{a vector, row or column of a matrix of initial values}

\item{targetValue}{a target value for the sum of the elements of the vector, row or column}
}
\value{
a vector whose sum is equal to the target value
}
\description{
A detailed explanation is available here: https://kamecon.quarto.pub/ras/#the-allocnp-function-figaro-implementation-of-gras
}
