% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_as_dt.R
\name{fig_as_dt}
\alias{fig_as_dt}
\title{Convert a vector or matrix to a data.table}
\usage{
fig_as_dt(input)
}
\arguments{
\item{input}{a vector or a matrix}
}
\value{
a data.table
}
\description{
Convert a vector or matrix to a data.table
}
