% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_date_time.R
\name{fig_date_time}
\alias{fig_date_time}
\title{Create a time and date stamp for Figaro output files}
\usage{
fig_date_time()
}
\value{
a formatted time stamp
}
\description{
Create a time and date stamp for Figaro output files
}
