% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_extract_matrices_figaro.R
\name{fig_extract_matrices_figaro}
\alias{fig_extract_matrices_figaro}
\title{Extract FIGARO table}
\usage{
fig_extract_matrices_figaro(
  path = "https://ec.europa.eu/eurostat/documents/51957/19580762/",
  year = NULL,
  quiet = FALSE
)
}
\arguments{
\item{path}{string}

\item{year}{Integer}

\item{quiet}{Boolean, if `TRUE`, suppress all status messages. Default
is `FALSE`, i.e., messages are shown.}
}
\value{
List with basic input-output matrices and metadata
}
\description{
Extract basic matrices from source data and prepares it for
processing with `make_wio`
}
