% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_extract_matrices_oecd_icio.R
\name{fig_extract_matrices_oecd_icio}
\alias{fig_extract_matrices_oecd_icio}
\title{Extract ICIO table}
\usage{
fig_extract_matrices_oecd_icio(path, year = NULL, quiet = FALSE)
}
\arguments{
\item{path}{Character string for source folder of zip file}

\item{year}{Integer}

\item{quiet}{Boolean, if `TRUE`, suppress all status messages. Default
is `FALSE`, i.e., messages are shown.}
}
\value{
List with basic input-output matrices and metadata
#' @export fig_extract_matrices_oecd_icio
}
\description{
Extract basic matrices from source data and prepares it for
processing with `make_wio`
}
