% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_figaro_countries.R
\docType{data}
\name{fig_figaro_countries}
\alias{fig_figaro_countries}
\title{Code list for Figaro countries (including 2025 new ones)}
\format{
An object of class \code{character} of length 49.
}
\usage{
fig_figaro_countries
}
\description{
Code list for Figaro countries (including 2025 new ones)
}
\keyword{datasets}
