% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_from_csv.R
\name{fig_from_csv}
\alias{fig_from_csv}
\title{Read csv files with right datatable options for NA}
\usage{
fig_from_csv(file, ...)
}
\arguments{
\item{file}{file to read}

\item{...}{other additional parameters}
}
\value{
a data.table
}
\description{
Read csv files with right datatable options for NA
}
