% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_get_icio.R
\name{fig_get_icio}
\alias{fig_get_icio}
\title{Prepares a virtual dataset of parquet files to extract ICIO data
You can apply additional filters and do not forget to collect the data}
\usage{
fig_get_icio(
  folder = "E:/users/biedmlu/figaro_db/24ed/reference",
  ic = "icsup"
)
}
\arguments{
\item{folder}{Folder where the files are. In the server now: "E:/users/biedmlu/figaro_db/24ed/reference"}

\item{ic}{the type to query ("icsup","icuse","iciop","icioi"). By default "icsup"}
}
\value{
an arrow virtual dataset
}
\description{
Prepares a virtual dataset of parquet files to extract ICIO data
You can apply additional filters and do not forget to collect the data
}
\examples{
tmp <- fig_get_icio() |>
filter(REF_AREA == "ES" & OBS_VALUE !=0 & COUNTERPART_AREA==REF_AREA) |>
collect()
}
