% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_get_nama.R
\name{fig_get_nama}
\alias{fig_get_nama}
\title{Extract NAMA data from MDT
Some columns are ignored and some filters applied to make it faster (only current prices). For very particular extractions use the normal procedure.}
\usage{
fig_get_nama(db_nama = "NAMA_REFERENCE_25ED", year_nama = seq(2010, 2023, 1L))
}
\arguments{
\item{db_nama}{MDT table to query, default option is "NAMA_REFERENCE_25ED"}

\item{year_nama}{years to extract, by defaul is 2010:2023}
}
\value{
a data.table
}
\description{
Extract NAMA data from MDT
Some columns are ignored and some filters applied to make it faster (only current prices). For very particular extractions use the normal procedure.
}
\examples{
dt <- fig_get_nama(db_nama = "NAMA_PRODUCTION_25ED", year_nama = 2023)
}
