% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_get_sut.R
\name{fig_get_sut}
\alias{fig_get_sut}
\title{Extract SUT data from MDT
Some columns are ignored and some filters applied to make it faster (only current prices). For very particular extractions use the normal procedure.}
\usage{
fig_get_sut(
  db_sut = "SUT_REFERENCE_25ED",
  year_sut = seq(2010, 2023, 1L),
  tables_sut = c("T1500", "T1600", "T1611", "T1612", "T1620", "T1630"),
  geo_sut
)
}
\arguments{
\item{db_sut}{MDT table to query, default option is "SUT_REFERENCE_25ED"}

\item{year_sut}{years to extract, by defaul is 2010:2023}

\item{tables_sut}{tables to extract by default c("T1500","T1600","T1611","T1612","T1620","T1630")}

\item{geo_sut}{countries to extract. A value is required}
}
\value{
a data.table
}
\description{
Extract SUT data from MDT
Some columns are ignored and some filters applied to make it faster (only current prices). For very particular extractions use the normal procedure.
}
\examples{
dt <- fig_get_sut(db_sut = "SUT_PRODUCTION_25ED", year_sut = 2021, geo_sut = "ES")
}
