% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_GRAS.R
\name{fig_GRAS}
\alias{fig_GRAS}
\title{Apply a standard 2 dimension GRAS}
\usage{
fig_GRAS(
  mPrior,
  cTarget,
  rTarget,
  balancingError = 1e-10,
  maxIter = 500,
  verbose = TRUE,
  printMatrix = FALSE
)
}
\arguments{
\item{mPrior}{a matrix with the prior values}

\item{cTarget}{Column target}

\item{rTarget}{Row target}

\item{balancingError}{Tolerance factor (default=0.001)}

\item{maxIter}{Maximum number of iterations (default=500)}

\item{verbose}{Print iteration number and error at each iteration (default=TRUE)}

\item{printMatrix}{Print matrix at each iteration (default=FALSE)}
}
\value{
a matrix satisfiying that rowsums and colsums equals the row and column targets
}
\description{
Apply a standard 2 dimension GRAS
}
\examples{
#
}
