% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_industry64.R
\docType{data}
\name{fig_industry64}
\alias{fig_industry64}
\title{Code list for 64 Figaro activities}
\format{
An object of class \code{character} of length 64.
}
\usage{
fig_industry64
}
\description{
Code list for 64 Figaro activities
}
\keyword{datasets}
