% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_iso.R
\docType{data}
\name{fig_iso}
\alias{fig_iso}
\title{Labels, ISO 2, ISO 3 and numeric codes for countries}
\format{
An object of class \code{data.frame} with 250 rows and 4 columns.
}
\usage{
fig_iso
}
\description{
Labels, ISO 2, ISO 3 and numeric codes for countries
}
\keyword{datasets}
