% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_product64.R
\docType{data}
\name{fig_product64}
\alias{fig_product64}
\title{Code list for 64 Figaro products}
\format{
An object of class \code{character} of length 64.
}
\usage{
fig_product64
}
\description{
Code list for 64 Figaro products
}
\keyword{datasets}
