% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_row_pi_dic.R
\docType{data}
\name{fig_row_pi_dic}
\alias{fig_row_pi_dic}
\title{Code list and labels for Figaro rowPi}
\format{
An object of class \code{data.frame} with 102 rows and 2 columns.
}
\usage{
fig_row_pi_dic
}
\description{
Code list and labels for Figaro rowPi
}
\keyword{datasets}
