% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_to_csv.R
\name{fig_to_csv}
\alias{fig_to_csv}
\title{Write csv files with right datatable options for dates}
\usage{
fig_to_csv(data, ...)
}
\arguments{
\item{data}{a data frame/data.table/tibble}

\item{...}{destination and other optional parameters}
}
\value{
a file in the destination provided
}
\description{
Write csv files with right datatable options for dates
}
\examples{
fig_to_csv(mtcars,"D:/test.csv")
}
