% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_usd_eur.R
\docType{data}
\name{fig_usd_eur}
\alias{fig_usd_eur}
\title{Exchange rate to convert USD to EUR (divide)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.
}
\usage{
fig_usd_eur
}
\description{
Exchange rate to convert USD to EUR (divide)
}
\keyword{datasets}
