% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_world_countries.R
\docType{data}
\name{fig_world_countries}
\alias{fig_world_countries}
\title{Code list for 202 world countries according to 2022-2024 figaro}
\format{
An object of class \code{character} of length 205.
}
\usage{
fig_world_countries
}
\description{
Code list for 202 world countries according to 2022-2024 figaro
}
\keyword{datasets}
