% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figutils-package.R
\docType{package}
\name{figutils-package}
\alias{figutils}
\alias{figutils-package}
\title{figutils: Set of utilities for interactive use of Figaro}
\description{
Collection of functions, dictionaries and other utilities to work interactively with Figaro data. It could be used also to retrieve easily common functions, parameters, etc used in Figaro to be pasted in production code instead of looking for them in other scripts.
}
\keyword{internal}
